Moved project files around, renamed projects and gave them new guids, turned the sample app back into an AppX package
