+++ Samples/Islands/Lottie/LottieIsland/packages.config
