+++ Samples/Islands/Lottie/SimpleLottieIslandApp.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34018.315
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SimpleIslandApp", "SimpleIslandApp.vcxproj", "{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SimpleLottieIslandApp", "SimpleLottieIslandApp\SimpleLottieIslandApp.vcxproj", "{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LottieIsland", "..\LottieIsland\LottieIsland.vcxproj", "{9BA7A2F2-B723-458B-A575-3F726D271465}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LottieIsland", "LottieIsland\LottieIsland.vcxproj", "{9BA7A2F2-B723-458B-A575-3F726D271465}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,31 +21,31 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|Any CPU.Build.0 = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|Any CPU.Deploy.0 = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|ARM.ActiveCfg = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|ARM.Build.0 = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|ARM64.Build.0 = Debug|ARM64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x64.ActiveCfg = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x64.Build.0 = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x64.Deploy.0 = Debug|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x86.ActiveCfg = Debug|Win32
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x86.Build.0 = Debug|Win32
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Debug|x86.Deploy.0 = Debug|Win32
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|Any CPU.ActiveCfg = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|Any CPU.Build.0 = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|Any CPU.Deploy.0 = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|ARM.ActiveCfg = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|ARM.Build.0 = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|ARM64.ActiveCfg = Release|ARM64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|ARM64.Build.0 = Release|ARM64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|x64.ActiveCfg = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|x64.Build.0 = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|x64.Deploy.0 = Release|x64
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|x86.ActiveCfg = Release|Win32
-		{350A5EC2-B156-4AAF-9D80-A864C76BA0C5}.Release|x86.Build.0 = Release|Win32
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|Any CPU.Build.0 = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|Any CPU.Deploy.0 = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|ARM.ActiveCfg = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|ARM.Build.0 = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|ARM64.Build.0 = Debug|ARM64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x64.ActiveCfg = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x64.Build.0 = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x64.Deploy.0 = Debug|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x86.ActiveCfg = Debug|Win32
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x86.Build.0 = Debug|Win32
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Debug|x86.Deploy.0 = Debug|Win32
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|Any CPU.ActiveCfg = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|Any CPU.Build.0 = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|Any CPU.Deploy.0 = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|ARM.ActiveCfg = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|ARM.Build.0 = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|ARM64.ActiveCfg = Release|ARM64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|ARM64.Build.0 = Release|ARM64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|x64.ActiveCfg = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|x64.Build.0 = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|x64.Deploy.0 = Release|x64
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|x86.ActiveCfg = Release|Win32
+		{44F9C1B5-DDDA-4D2E-8AD4-205A40E15192}.Release|x86.Build.0 = Release|Win32
 		{9BA7A2F2-B723-458B-A575-3F726D271465}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{9BA7A2F2-B723-458B-A575-3F726D271465}.Debug|Any CPU.Build.0 = Debug|x64
 		{9BA7A2F2-B723-458B-A575-3F726D271465}.Debug|ARM.ActiveCfg = Debug|ARM
