+++ Samples/Islands/Lottie/SimpleLottieIslandApp/AnimatedVisuals.LottieLogo1.idl
