+++ Samples/Islands/Lottie/SimpleLottieIslandApp/Resource.h
@@ -7,13 +7,13 @@
 #define IDS_APP_TITLE			103
 
 #define IDR_MAINFRAME			128
-#define IDD_SIMPLEISLANDAPP_DIALOG	102
+#define IDD_SIMPLELOTTIEISLANDAPP_DIALOG	102
 #define IDD_ABOUTBOX			103
 #define IDM_ABOUT				104
 #define IDM_EXIT				105
-#define IDI_SIMPLEISLANDAPP			107
+#define IDI_SIMPLELOTTIEISLANDAPP			107
 #define IDI_SMALL				108
-#define IDC_SIMPLEISLANDAPP			109
+#define IDC_SIMPLELOTTIEISLANDAPP			109
 #define IDC_MYICON				2
 #ifndef IDC_STATIC
 #define IDC_STATIC				-1
