+++ Samples/Islands/Lottie/SimpleLottieIslandApp/SimpleLottieIslandApp.ico
