+++ Samples/Islands/Lottie/SimpleLottieIslandApp/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Graphics.Win2D" version="1.2.0" targetFramework="native" />
   <package id="Microsoft.Windows.CppWinRT" version="2.0.230706.1" targetFramework="native" />
   <package id="Microsoft.Windows.SDK.BuildTools" version="10.0.22621.1" targetFramework="native" />
   <package id="Microsoft.WindowsAppSDK" version="1.5.240227000" targetFramework="native" />
