+++ Samples/Islands/Lottie/SimpleLottieIslandApp/small.ico
