+++ Samples/Islands/Lottie/nuget.config
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <config>
+        <add key="globalPackagesFolder" value=".\packages" />
+    </config>
+</configuration>
